Finished quiz and fixed course card layout
