+++ lib/views/course/course.dart
@@ -2,10 +2,9 @@ import 'package:bro/views/course/quiz.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:graphql_flutter/graphql_flutter.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:flutter/foundation.dart';
-import 'info_card.dart';
-import 'package:dots_indicator/dots_indicator.dart';
+import 'course_card_view.dart';
 import 'package:bro/blocs/course_detail/course_detail_bucket.dart';
 import 'package:bro/models/course.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -58,12 +57,16 @@ class _CourseDetailViewState extends State<CourseDetailView> {
           data = state.course;
           if (!state.isQuiz) {
             return Scaffold(
-              appBar: AppBar(title: Text(data.title)),
+              appBar: AppBar(
+                  title: Text(data.title),
+                  leading: _ExitVerification(context, data)),
               body: _course_view_builder(context, data),
             );
           } else {
             return Scaffold(
-              appBar: AppBar(title: Text(data.title)),
+              appBar: AppBar(
+                  title: Text(data.title),
+                  leading: _ExitVerification(context, data)),
               body: Center(
                   child: QuizView(
                       course: data,
@@ -80,162 +83,58 @@ class _CourseDetailViewState extends State<CourseDetailView> {
   }
 }
 
+Widget _ExitVerification(context, data) {
+  return IconButton(
+      icon: Icon(Icons.arrow_back, color: Colors.teal),
+      onPressed: () => {
+            showDialog(
+                context: context,
+                builder: (_) => AlertDialog(
+                    titlePadding: EdgeInsets.all(2),
+                    contentPadding: EdgeInsets.fromLTRB(25, 0, 25, 25),
+                    title: Align(
+                      alignment: Alignment.topRight,
+                      child: GestureDetector(
+                        child: FaIcon(FontAwesomeIcons.solidWindowClose,
+                            color: Colors.red[900], size: 22),
+                        onTap: () => Navigator.of(context).pop(),
+                      ),
+                    ),
+                    content: Column(mainAxisSize: MainAxisSize.min, children: [
+                      Container(
+                        alignment: Alignment.bottomCenter,
+                        padding: EdgeInsets.only(bottom: 25),
+                        child: Text('Avbryt kurs',
+                            textAlign: TextAlign.center,
+                            style: TextStyle(color: Colors.teal, fontSize: 22)),
+                      ),
+                      Text(
+                        'Er du sikker på at du vil avbryte kurset?\n\nVed å gjøre dette lagres ikke progresjonen din.',
+                        textAlign: TextAlign.center,
+                      ),
+                      Padding(
+                        padding: const EdgeInsets.only(top: 20),
+                        child: ElevatedButton(
+                          child: Text(
+                            '        Avbryt Kurs        ',
+                          ),
+                          onPressed: () => {
+                            Navigator.of(context)..pop()..pop(),
+                          },
+                        ),
+                      ),
+                    ]))),
+          });
+}
+
 Widget _course_view_builder(context, data) {
   return Container(
+      padding: EdgeInsets.only(top: MediaQuery.of(context).size.height * 0.1),
+      alignment: Alignment.center,
       width: MediaQuery.of(context).size.width,
-      height: MediaQuery.of(context).size.height * 0.7,
+      height: MediaQuery.of(context).size.height * 0.9,
       child: CardContainerView(
         list: data.slides,
         course: data,
       ));
-}
-
-class CardContainerView extends StatefulWidget {
-  CardContainerView({
-    Key key,
-    this.list,
-    this.res,
-    this.course,
-  }) : super(key: key);
-  final List list;
-  final QueryResult res;
-  final Course course;
-  @override
-  _CardContainerViewState createState() => _CardContainerViewState();
-}
-
-class _CardContainerViewState extends State<CardContainerView> {
-  ScrollController _controller;
-  double indx = 0;
-  @override
-  void initState() {
-    _controller = ScrollController();
-    _controller.addListener(() {
-      if (indx != (_controller.offset / context.size.width).round()) {
-        setState(() {
-          indx = (_controller.offset / context.size.width).roundToDouble();
-        });
-      }
-      print(indx);
-      //SET LIMITER PÅ ANTALL UTREGNINGER?
-    });
-    super.initState();
-  }
-
-//ScrollController functions for swiping right and left
-  void _moveRight() {
-    _controller
-        .animateTo(_controller.offset + context.size.width,
-            curve: Curves.linear, duration: Duration(milliseconds: 200))
-        .whenComplete(() => setState(() {
-              indx = _controller.offset / context.size.width;
-            }));
-  }
-
-  void _moveLeft() {
-    _controller
-        .animateTo(_controller.offset - context.size.width,
-            curve: Curves.linear, duration: Duration(milliseconds: 200))
-        .whenComplete(() => setState(() {
-              indx = _controller.offset / context.size.width;
-            }));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (context == null) {
-      return Scaffold(body: Text('Context is null, yo'));
-    }
-
-    if (widget.list.isNotEmpty) {
-      return Column(children: [
-        Container(
-            width: MediaQuery.of(context).size.width,
-            height: MediaQuery.of(context).size.height * 0.5,
-            child: ListView.builder(
-                shrinkWrap: true,
-                controller: _controller,
-                scrollDirection: Axis.horizontal,
-                physics: PageScrollPhysics(),
-                itemCount: widget.list.length,
-                itemBuilder: (BuildContext context, int index) {
-                  return InfoCard(
-                      title: widget.list[index]['title'],
-                      description: widget.list[index]['description'],
-                      image: widget.list[index]['image']);
-                })),
-        //Shows which page the user is on
-        DotsIndicator(
-            dotsCount: widget.list.length,
-            position: indx.roundToDouble(),
-            decorator: DotsDecorator(
-                color: Colors.grey[350], activeColor: Colors.teal)),
-        //Scroll buttons
-        Container(
-          width: MediaQuery.of(context).size.width,
-          child: Row(
-              mainAxisAlignment: indx + 1 == widget.list.length
-                  ? MainAxisAlignment.start
-                  : MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                GestureDetector(
-                    onTap: () => {_moveLeft()},
-                    child: Align(
-                      alignment: Alignment.centerLeft,
-                      child: Padding(
-                          padding: EdgeInsets.only(
-                              left: MediaQuery.of(context).size.width * 0.073),
-                          child: RotatedBox(
-                              quarterTurns: 1,
-                              child: Icon(
-                                Icons.arrow_circle_down,
-                                color: Colors.teal,
-                                size: MediaQuery.of(context).size.width * 0.175,
-                              ))),
-                    )),
-                indx + 1 != widget.list.length
-                    ? GestureDetector(
-                        onTap: () => {_moveRight()},
-                        child: Align(
-                          alignment: Alignment.centerLeft,
-                          child: Padding(
-                              padding: EdgeInsets.symmetric(
-                                  vertical: 0, horizontal: 30),
-                              child: RotatedBox(
-                                  quarterTurns: 3,
-                                  child: Icon(
-                                    Icons.arrow_circle_down,
-                                    color: Colors.teal,
-                                    size: MediaQuery.of(context).size.width *
-                                        0.175,
-                                  ))),
-                        ))
-                    : Padding(
-                        padding: EdgeInsets.symmetric(
-                            horizontal:
-                                MediaQuery.of(context).size.width * 0.05),
-                        child: ButtonTheme(
-                          child: ElevatedButton(
-                            child: Padding(
-                              padding: const EdgeInsets.symmetric(
-                                  horizontal: 40, vertical: 15),
-                              child: Center(child: Text('Start Quiz')),
-                            ),
-                            onPressed: () {
-                              BlocProvider.of<CourseDetailBloc>(context).add(
-                                  CourseDetailRequested(
-                                      course: widget.course,
-                                      isQuiz: true,
-                                      isAnswer: false));
-                            },
-                          ),
-                        ),
-                      )
-              ]),
-        )
-      ]);
-    } else {
-      return Container(width: 0.0, height: 0.0);
-    }
-  }
 }
